requirements.txt
numpy==1.18.1
matplotlib==3.1.3
torchvision==0.7.0
Pillow==8.0.1
torch==1.7.0

